requirements.txt
Flask 
flask_cors 
gnupg 
PGPy
dnspython

